+++ CHANGELOG.md
@@ -1,13 +1,28 @@
-yumrepo v0.16.3
+## yumrepo
+
+### v0.16.4
+
+Thanks to Paul Graydon (@twirrim) for adding Percona MySQL repositories.
+
+Thanks to Bryan Berry (@bryanwb) for adding yet another PostgreSQL
+repository. :)
+
+#### Breaking changes (postgresql)
+
+* Recipe renamed, postgresql9 -> postgresql.
+* Added node['repo']['postgresql']['version'] to select version of
+  Postgres via attribute, from pgrpms.org repository.
+
+### v0.16.3
 
 Set name/base_url to localdomain in the case Ohai `domain` is
 unavailable.  Fixes chef-solo compilation error on Vagrant boxes.
 
-## v0.16.2
+### v0.16.2
 
     Change corporate recipe to key off domain name.
 
-## v0.16.0
+### v0.16.0
 
     Add VMWare Tools for ESX v5.0u1, along with a vmware-tools-upgrade
 recipe.  The yumrepo::vmware-tools recipe is backwards incompatible.